+++ src/vue-router/index.js
@@ -30,7 +30,8 @@ import AtomicSwaps from '@/vue/pages/AtomicSwaps'
 import AtomicSwapsExplore from '@/vue/pages/atomic-swaps/AtomicSwapsExplore'
 import Settings from '@/vue/pages/Settings'
 import Verification from '@/vue/pages/Verification'
-import VerificationFormContainer from '@/vue/pages/VerificationFormContainer'
+import VerificationGeneral from '@/vue/pages/VerificationGeneral'
+import VerificationCorporate from '@/vue/pages/VerificationCorporate'
 import Security from '@/vue/pages/Security'
 import AssetExplorerModule from '@modules/assets/asset-explorer'
 import BalanceExplorerModule from '@modules/assets/balance-explorer'
@@ -396,13 +397,13 @@ const router = new Router({
                 {
                   path: '/settings/verification/general',
                   name: vueRoutes.verificationGeneral.name,
-                  component: VerificationFormContainer,
+                  component: VerificationGeneral,
                   beforeEnter: inAppRouteGuard,
                 },
                 {
                   path: '/settings/verification/corporate',
                   name: vueRoutes.verificationCorporate.name,
-                  component: VerificationFormContainer,
+                  component: VerificationCorporate,
                   beforeEnter: inAppRouteGuard,
                 },
               ],
