+++ src/vue/pages/verification/styles.scss
@@ -1,0 +1,15 @@
+@import '~@/vue/forms/_app-form';
+
+.verification-container__heading {
+  font-size: 1.3rem;
+  font-weight: 400;
+  margin-top: 4rem;
+}
+
+.verification-container__form {
+  margin-top: 1rem;
+  background-color: $col-block-bg;
+  padding: 2.4rem;
+
+  @include box-shadow();
+}
